1.0.0.6
Exploring the children and find methods
children is used to access the direct children elements. This also
call the css method of calling for example "ul > li". The ">" inbetween
denotes the direct children access.
find is used to access all decendent of an element and "ul li" (example)
can be used.