1.0.4.35
Created a problem scenario where a value is assigned to a variable within a timeout (2 secs). outside of this function. this
variable value is unknown as we may not know who long will it take for a value to be assigned.

In the next example we will use the defferds object  to solve this